Minor refactor for the new code implementing PATCH /entities/{entityId}
