+++ src/lib/orionld/db/dbModelToApiSubAttribute.cpp
@@ -1,0 +1,58 @@
+/*
+*
+* Copyright 2022 FIWARE Foundation e.V.
+*
+* This file is part of Orion-LD Context Broker.
+*
+* Orion-LD Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion-LD Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* orionld at fiware dot org
+*
+* Author: Ken Zangelin
+*/
+extern "C"
+{
+#include "kbase/kMacros.h"                                       // K_VEC_SIZE
+#include "kjson/KjNode.h"                                        // KjNode
+#include "kjson/kjLookup.h"                                      // kjLookup
+#include "kjson/kjBuilder.h"                                     // kjChildRemove
+}
+
+#include "logMsg/logMsg.h"                                       // LM_*
+
+#include "orionld/common/orionldState.h"                         // orionldState
+#include "orionld/db/dbModelToApiSubAttribute.h"                 // Own interface
+
+
+
+// -----------------------------------------------------------------------------
+//
+// dbModelToApiSubAttribute -
+//
+void dbModelToApiSubAttribute(KjNode* subP)
+{
+  //
+  // Remove unwanted parts of the sub-attribute from DB
+  //
+  const char* unwanted[] = { "createdAt", "modifiedAt" };
+
+  for (unsigned int ix = 0; ix < K_VEC_SIZE(unwanted); ix++)
+  {
+    KjNode* nodeP = kjLookup(subP, unwanted[ix]);
+
+    if (nodeP != NULL)
+      kjChildRemove(subP, nodeP);
+  }
+}
