+++ src/lib/orionld/serviceRoutines/orionldPatchEntity2.cpp
@@ -41,7 +41,6 @@ extern "C"
 #include "orionld/common/orionldError.h"                         // orionldError
 #include "orionld/common/dotForEq.h"                             // dotForEq
 #include "orionld/common/eqForDot.h"                             // eqForDot
-#include "orionld/common/numberToDate.h"                         // numberToDate
 #include "orionld/types/OrionldHeader.h"                         // orionldHeaderAdd
 #include "orionld/kjTree/kjJsonldNullObject.h"                   // kjJsonldNullObject
 #include "orionld/kjTree/kjTreeLog.h"                            // kjTreeLog
@@ -50,45 +49,38 @@ extern "C"
 #include "orionld/kjTree/kjStringValueLookupInArray.h"           // kjStringValueLookupInArray
 #include "orionld/payloadCheck/pCheckEntity.h"                   // pCheckEntity
 #include "orionld/db/dbModelFromApiEntity.h"                     // dbModelFromApiEntity
+#include "orionld/db/dbModelToApiAttribute.h"                    // dbModelToApiAttribute
 #include "orionld/serviceRoutines/orionldPatchEntity2.h"         // Own Interface
 
 
 
-//
-// All db functions (all new ones anyway) should return an entity from DB AS IS, i.e. include the database model
-// Then we need functions to adapt to and from DB-Model:
-//
-// dbModelToApiEntity           - Strip Entity of extra DB model stuff and produce an NGSI-LD API Entity (attrNames is "hidden" as .names)
-// dbModelToApiAttribute        - ... produce an NGSI-LD API Attribute (mdNames is "hidden as .names)
-// dbModelToApiSubAttribute     - ... produce an NGSI-LD API Sub-Attribute
-// dbModelFromApiEntity         - convert an NGSI-LD API Entity into an "Orion DB" Entity
-// dbModelFromApiAttribute      -
-// dbModelFromApiSubAttribute   -
 //
 // To PATCH an Entity:
 // 1. GET the entity from DB - in DB-Model style
 // 2. Convert the request payload to a DB-Model style Entity (includes modifiedAt but not createdAt) - dbModelFromApiXXX()
 // 3. Compare the two trees and come up with an array of changes (orionldEntityPatchTree - output from orionldPatchXXX functions):
 //    [
-//      { PATH, TREE },
-//      { PATH, TREE },
+//      { PATH, TREE, op },
+//      { PATH, TREE, op },
 //      ...
 //    ]
 //
-//    E.g. (modification of a compound value of a sub-attr   +  addition of an attribute + removal of a sub-attr):
+//    E.g. (P1: modification of a compound value of a sub-attr   +  P2: addition of an attribute + P3: removal of a sub-attr):
 //    [
 //      {
-//        "path": "attrs.P1.md.Sub-R.value.F",
-//        "tree": <KjNode tree of the new value>
+//        "PATH": "attrs.P1.md.Sub-R.value.F",
+//        "TREE": <KjNode tree of the new value>,
+//        "op": "Modify" - Not needed - that is the default
 //      },
 //      {
-//        "path": "attrs.P2",
-//        "tree": <KjNode tree of the attribute - DB Model style>,
+//        "PATH": "attrs.P2",
+//        "TREE": <KjNode tree of the attribute - DB Model style>,
+//        "op": "Add" - Not needed - "modify" and "add" is the same for mongo - REPLACE/APPEND
 //      },
 //      {
-//        "path": [ "attrs", "P3", "md", "SP1" ]
-//        "tree": null,
-//        "op": "delete"
+//        "PATH": "attrs.P3.md.SP1",
+//        "TREE": null,
+//        "op": "delete"  - Not needed - "TREE" is null, so, the Delete info is already present
 //      }
 //    ]
 //
@@ -105,7 +97,7 @@ extern "C"
 //
 // patchTreeItemAdd -
 //
-void patchTreeItemAdd(KjNode* patchTree, const char* path, KjNode* valueP, const char* op)
+static void patchTreeItemAdd(KjNode* patchTree, const char* path, KjNode* valueP, const char* op)
 {
   KjNode* arrayItemP = kjObject(orionldState.kjsonP, NULL);
   KjNode* pathP      = kjString(orionldState.kjsonP, "PATH", path);
@@ -128,13 +120,14 @@ void patchTreeItemAdd(KjNode* patchTree, const char* path, KjNode* valueP, const
 
 // -----------------------------------------------------------------------------
 //
-// namesArrayMergeToPatchTree -
+// namesArrayMergeToPatchTree - only used if we both add and delete attrs in the same operation
 //
 static void namesArrayMergeToPatchTree(KjNode* patchTree, const char* path, KjNode* namesP, KjNode* addedP, KjNode* removedP)
 {
   for (KjNode* rmP = removedP->value.firstChildP; rmP != NULL; rmP = rmP->next)
   {
     KjNode* itemP = kjStringValueLookupInArray(namesP, rmP->value.s);
+
     if (itemP)
       kjChildRemove(namesP, itemP);
   }
@@ -199,7 +192,7 @@ static void namesToPatchTree(KjNode* patchTree, const char* path, KjNode* namesP
 //
 // Might be a good idea to sort both trees before starting to process ...
 //
-void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patchTree)
+static void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patchTree)
 {
   if (newP == NULL)  // Not sure this is ever a possibility ...
     return;
@@ -259,15 +252,18 @@ void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patc
     return;
   }
 
-  // Both are JSON Object - entering to find differences
+  // Both newP and oldP are JSON OBJECT - entering inside the objects to find differences
 
-  //
-  // Careful with the linked lists of newP - the recursive calls may invoke patchTreeItemAdd and that REMOVES ITEMS FROM A LIST
-  // => Can't use the normal "for" - must be a "while and a next pointer"
-  //
   KjNode* namesP     = kjLookup(newP, ".names");
   KjNode* addedP     = kjLookup(newP, ".added");
   KjNode* removedP   = kjLookup(newP, ".removed");
+
+
+
+  //
+  // CAREFUL with the linked lists of newP - the recursive calls may invoke patchTreeItemAdd and that *REMOVES ITEMS FROM A LIST*
+  // => Can't use the normal "for" - must be a "while and a next pointer"
+  //
   KjNode* newItemP   = newP->value.firstChildP;
   KjNode* next;
 
@@ -307,93 +303,6 @@ void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patc
 
 
 
-void dbModelToApiSubAttribute(KjNode* subP)
-{
-  //
-  // Remove unwanted parts of the sub-attribute from DB
-  //
-  const char* unwanted[] = { "createdAt", "modifiedAt" };
-
-  for (unsigned int ix = 0; ix < K_VEC_SIZE(unwanted); ix++)
-  {
-    KjNode* nodeP = kjLookup(subP, unwanted[ix]);
-
-    if (nodeP != NULL)
-      kjChildRemove(subP, nodeP);
-  }
-}
-
-
-
-void dbModelToApiAttribute(KjNode* attrP)
-{
-  //
-  // Remove unwanted parts of the attribute from DB
-  //
-  const char* unwanted[] = { "mdNames", "creDate", "modDate" };
-
-  for (unsigned int ix = 0; ix < K_VEC_SIZE(unwanted); ix++)
-  {
-    KjNode* nodeP = kjLookup(attrP, unwanted[ix]);
-
-    if (nodeP != NULL)
-      kjChildRemove(attrP, nodeP);
-  }
-
-  KjNode* observedAtP = kjLookup(attrP, "observedAt");
-  KjNode* unitCodeP   = kjLookup(attrP, "unitCode");
-
-  if (observedAtP != NULL)
-  {
-    char* dateTimeBuf = kaAlloc(&orionldState.kalloc, 32);
-    numberToDate(observedAtP->value.firstChildP->value.f, dateTimeBuf, 32);
-    observedAtP->type      = KjString;
-    observedAtP->value.s   = dateTimeBuf;
-    observedAtP->lastChild = NULL;
-  }
-
-  if (unitCodeP != NULL)
-  {
-    unitCodeP->type      = KjString;
-    unitCodeP->value.s   = unitCodeP->value.firstChildP->value.s;
-    unitCodeP->lastChild = NULL;
-  }
-
-  //
-  // Sub-Attributes
-  //
-  KjNode* mdP = kjLookup(attrP, "md");
-  if (mdP != NULL)
-  {
-    kjChildRemove(attrP, mdP);  // The content of mdP is added to attrP at the end of the if
-
-    // Special Sub-Attrs: unitCode + observedAt
-    for (KjNode* subP = mdP->value.firstChildP; subP != NULL; subP = subP->next)
-    {
-      if (strcmp(subP->name, "unitCode") == 0)
-      {
-        subP->type  = subP->value.firstChildP->type;
-        subP->value = subP->value.firstChildP->value;
-      }
-      else if (strcmp(subP->name, "observedAt") == 0)  // Part of Sub-Attribute
-      {
-        subP->type  = subP->value.firstChildP->type;
-        subP->value = subP->value.firstChildP->value;  // + convert to iso8601 string
-      }
-      else
-        dbModelToApiSubAttribute(subP);
-    }
-
-    //
-    // Move all metadata (sub-attrs) up one level
-    //
-    attrP->lastChild->next = mdP->value.firstChildP;
-    attrP->lastChild       = mdP->lastChild;
-  }
-}
-
-
-
 // ----------------------------------------------------------------------------
 //
 // orionldPatchEntity2 -
@@ -418,7 +327,11 @@ bool orionldPatchEntity2(void)
   //        * create a KjNode array with all toplevel field names
   //        * sort out non-attribute names (id, type, scope, ...)
   //
-  //        Perhaps later ...
+  //        All that is done by pCheckEntity, but pCheckEntity needs the DB Entity :(
+  //        The chicken and the egg ...
+  //
+  //        Not really worth it.
+  //        At least, postponed (not forgotten) for now.
   //
   dbEntityP = mongocEntityLookup(entityId);
